requirements.txt

aiohttp==3.8.4
Flask>=2.3.2
numpy==1.24.2
Pillow>= 1.0 # Python Image Library (PIL) for processing images
pytest==7.3.1
pytest-mock==3.10.0
requests==2.30.0
urllib3==2.0.2